server: ignore muted/silent source channels when writing to output buffer
This change is done to avoid bad audio quality if a stream is silent and there is other non-silent audio playing. For example, 4 source channels with two silent used to cause the 2 channels with data to all be focused on one output channel, which can be undesirable in many cases.